+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning], with the exception that minor rel
 
 - 🐛 Fix bug in MLIR roundtrip passes caused by accessing an invalidated iterator after erasure in a loop ([#932]) ([**@flowerthrower**])
 - 🐛 Add missing support for `sxdg` gates in Qiskit circuit import ([#930]) ([**@burgholzer**])
+- 🐛 Fix bug initialization of operations with duplicate operands ([#964]) ([**@ystade**])
 
 ### Changed
 
@@ -107,6 +108,7 @@ _📚 Refer to the [GitHub Release Notes](https://github.com/munich-quantum-tool
 
 <!-- PR links -->
 
+[#964]: https://github.com/munich-quantum-toolkit/core/pull/964
 [#934]: https://github.com/munich-quantum-toolkit/core/pull/934
 [#933]: https://github.com/munich-quantum-toolkit/core/pull/933
 [#932]: https://github.com/munich-quantum-toolkit/core/pull/932